Further work on arbitrary optimizator interface
