+++ epde/optimizers/single_criterion/ea_stop_conds.py
@@ -8,7 +8,7 @@ Created on Tue Jun 15 20:51:55 2021
 
 from abc import ABC, abstractproperty
 
-class Stop_condition(ABC):
+class StopCondition(ABC):
     '''
     
     Abstract class for generalized stop condition for the evolutionary algorithm.
@@ -31,10 +31,9 @@ class Stop_condition(ABC):
     def check(self):
         pass
     
-class Iteration_limit(Stop_condition):
-    def __init__(self, limit = 100): # estrategy, 
+class IterationLimit(StopCondition):
+    def __init__(self, limit = 100):
         self.limit = limit + 1
-#        self.estrategy = estrategy
         self.checks = 0 # shortcut^ checking a number of calls instead of the estrategy attr of some sort
         
     def reset(self):
@@ -44,7 +43,7 @@ class Iteration_limit(Stop_condition):
         self.checks += 1
         return self.checks > self.limit
     
-class Fitness_stabilized(Stop_condition):
+class FitnessStabilized(StopCondition):
     def __init__(self, estrategy):
         raise NotImplementedError
